requirements.txt
Django==1.8.9
argparse==1.2.1

wsgiref==0.1.2
google-api-python-client==1.5.4
python-social-auth==0.2.21
djangorestframework==3.5.2
Markdown==2.6.7
PyJWT==1.4.2
django-filter==0.15.3
httplib2==0.9.2
oauth2client==4.0.0
oauthlib==2.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
wsgiref==0.1.2
